test+fix(100-01): UltimateAccessControl hardening findings 1-205 (100 tests, 15 files)
- CRITICAL #93: lock(this) -> private _statsLock in AccessControlStrategyBase
- Naming: ComplianceStandard SOX->Sox, FederationProtocol OIDC->Oidc, SmartCardType PIV->Piv
- Naming: Fido2 isES256->isEs256, PredictiveThreat sumXY->sumXy, Casbin object_->@object
- Fix: PossibleLossOfFraction in DctCoefficient and LsbEmbedding strategies
- Fix: Culture-specific IndexOf in SteganographyStrategy
- Fix: Identical ternary branches in Fido2Strategy
- 100 hardening tests covering all 205 findings

+++ Plugins/DataWarehouse.Plugins.UltimateAccessControl/Strategies/Steganography/DctCoefficientHidingStrategy.cs
@@ -246,7 +246,7 @@ namespace DataWarehouse.Plugins.UltimateAccessControl.Strategies.Steganography
                 RawCapacityBits = rawCapacityBits,
                 UsableCapacityBits = usableBits,
                 UsableCapacityBytes = (int)(usableBits / 8),
-                EffectiveCapacityBytes = (int)((usableBits / 8) / efficiencyFactor),
+                EffectiveCapacityBytes = (int)((double)usableBits / 8 / efficiencyFactor),
                 BlockCount = structure.BlockCount,
                 UsableCoefficientsPerBlock = structure.UsableCoefficientsPerBlock,
                 EstimatedQuality = structure.EstimatedQuality
